Improvement of templates and views for import_export
