+++ reman/templates/reman/tab_select.html
@@ -1,10 +1,10 @@
 {% load i18n %}
 
-<li class="nav-item {% if select == 'import' %}active{% endif %}">
-    <a class="nav-link {% if select == 'import' %}active{% endif %}" href="{% url 'reman:import_export' %}">
-        {% trans "Import-Export" %}
-    </a>
-</li>
+{#<li class="nav-item {% if select == 'import' %}active{% endif %}">#}
+{#    <a class="nav-link {% if select == 'import' %}active{% endif %}" href="{% url 'reman:import_export' %}">#}
+{#        {% trans "Import-Export" %}#}
+{#    </a>#}
+{#</li>#}
 <li class="nav-item {% if select == 'batch' %}active{% endif %}">
     <a class="nav-link {% if select == 'batch' %}active{% endif %}" href="{% url 'reman:batch_table' %}">
         {% trans "batch list" %}
