+++ reman/tests/tests_reman_views.py
@@ -71,13 +71,3 @@ class RemanTestCase(UnitTest):
         self.login()
         response = self.client.get(reverse('reman:out_table'))
         self.assertEqual(response.status_code, 200)
-
-    def test_import_export_page(self):
-        response = self.client.get(reverse('reman:import_export'))
-        self.assertRedirects(response, '/accounts/login/?next=/reman/import-export/', status_code=302)
-
-        self.add_perms_user(EcuModel, 'add_ecumodel', 'change_ecumodel')
-        self.add_perms_user(Corvet, 'add_corvet')
-        self.login()
-        response = self.client.get(reverse('reman:import_export'))
-        self.assertEqual(response.status_code, 200)
