+++ templates/base.html
@@ -22,19 +22,7 @@
             <!-- Divider -->
             <hr class="sidebar-divider my-0">
 
-            <!-- Nav Item - Dashboard -->
-            <li class="nav-item">
-                <a class="nav-link" href="{% url 'dashboard:charts' %}">
-                    <i class=" fas fa-fw fa-tachometer-alt"></i>
-                    <span>{% trans "Dashboard" %}</span></a>
-            </li>
-
-            <!-- Nav Item - thermal chamber -->
-            <li class="nav-item">
-                <a class="nav-link" href="{% url 'tools:thermal' %}">
-                    <i class=" fas fa-fw fa-thermometer"></i>
-                    <span>{% trans "Thermal chamber" %}</span></a>
-            </li>
+            {% include 'nav/nav_general.html' %}
 
             <!-- Divider -->
             <hr class="sidebar-divider">
