+++ templates/nav/nav_csd_atelier.html
@@ -1,5 +1,4 @@
-{% load i18n %}
-{% load user_tags %}
+{% load i18n user_tags %}
 
 {% if perms.squalaetp or perms.reman or perms.raspeedi %}
