+++ templates/nav/nav_general.html
@@ -1,0 +1,22 @@
+{% load i18n %}
+
+<!-- Nav Item - Dashboard -->
+<li class="nav-item">
+    <a class="nav-link" href="{% url 'dashboard:charts' %}">
+        <i class=" fas fa-fw fa-tachometer-alt"></i>
+        <span>{% trans "Dashboard" %}</span></a>
+</li>
+
+<!-- Nav Item - Thermal Chamber -->
+<li class="nav-item">
+    <a class="nav-link" href="{% url 'tools:thermal' %}">
+        <i class=" fas fa-fw fa-thermometer"></i>
+        <span>{% trans "Thermal chamber" %}</span></a>
+</li>
+
+<!-- Nav Item - Import Export -->
+<li class="nav-item">
+    <a class="nav-link" href="{% url 'import_export:detail' %}">
+        <i class=" fas fa-fw fa-sync-alt"></i>
+        <span>{% trans "Import / Export" %}</span></a>
+</li>
