+++ templates/nav/nav_gestion.html
@@ -25,7 +25,6 @@
                     <h6 class="collapse-header">Administration :</h6>
                     <a class="batch-btn collapse-item" href="#">{% trans "Create Batch" %}</a>
                     <a class="collapse-item" href="{% url 'reman:batch_table' %}">{% trans "batch list" %}</a>
-                    <a class="collapse-item" href="{% url 'reman:import_export' %}">{% trans "Import-Export" %}</a>
                 </div>
             </div>
         </li>
